Create base issue filter for shared logic
